Remove use of base.in
We seemed to be replicating dependencies specified in pyproject.toml in
requirements/base.in. There appears to be little point in this (it just
makes updating versions harder). Was the only reason for this to be able
to have a dev.txt? But I think that this is not required there either,
since making a local install should also install those dependencies from
pyproject.toml.